Add FromRawFd trait implementation and modernize
The newly added `FromRawFd` function allows users to directly construct
a `Memfd` from a file descriptor without running the checks that
`Memfd::try_from_file` would. This is useful in cases where a cost of
such a check would be prohibitive and the user is willing to deal with
`unsafe`ty of the function.

In addition to this new API there are minor updates to the documentation
and removal of dependency on the heavy thiserror.