Change `add_seals` to take an iterator instead of a `&HashSet`.
Generalize `Memfd::add_seals` to take an iterator instead of a
`&HashSet`. This way, users can still pass in a `&HashSet` if they have
one, but they can also pass in a simple slice, which is more convenient.

This also updates the examples in examples/sized.rs and src/lib.rs to
use slices, which is easier to read. These changes aren't strictly
necessary, because the old code would still work.